requirements.txt
torch>=2.1
open_clip_torch>=2.24.0
faiss-cpu>=1.7.4
pandas>=2.0
numpy>=1.24
Pillow>=10.0
tqdm>=4.66
pyyaml>=6.0

rank_bm25>=0.2.2

scikit-learn>=1.4
joblib>=1.3

lightgbm==4.5.0 # optional; we use sklearn HGBT by default
# Use conda opencv in environment.yml; avoid duplicate pip install

# For intelligent frame sampling
decord>=0.6.0
scipy>=1.11.0

# NEW: Near-SOTA easy-to-implement alternatives
ultralytics>=8.0.0  # FastSAM (50x faster than SAM2)
easyocr>=1.7.0      # Best OCR for multilingual support
transformers>=4.30.0  # BLIP-2 image captioning
accelerate>=0.20.0    # For transformer models
pyarrow>=14.0.0
